+++ tests/tests_unit/rules/test_models/test_dms_architect_rules.py
@@ -21,6 +21,7 @@ from cognite.neat.rules.models.dms import (
     DMSViewInput,
 )
 from cognite.neat.utils.cdf_classes import ContainerApplyDict, NodeApplyDict, SpaceApplyDict, ViewApplyDict
+from tests.data import car
 
 
 def rules_schema_tests_cases() -> Iterable[ParameterSet]:
@@ -1550,6 +1551,20 @@ class TestDMSRules:
         actual_issues = validation.NeatValidationError.from_pydantic_errors(e.value.errors())
 
         assert sorted(actual_issues) == sorted(expected_issues)
+
+    def test_create_reference(self) -> None:
+        dms_rules = car.CAR_RULES.as_dms_architect_rules()
+
+        dms_rules.create_reference(car.BASE_MODEL, {"Manufacturer": "Entity", "Color": "Entity"})
+
+        schema = dms_rules.as_schema()
+        view_by_external_id = {view.external_id: view for view in schema.views.values()}
+        # The Manufacturer and Color view only has one property, name, and this is
+        # now expected to point to the Entity container in the base model.
+        manufacturer_view = view_by_external_id["Manufacturer"]
+        assert manufacturer_view.referenced_containers() == {dm.ContainerId(car.BASE_MODEL.metadata.space, "Entity")}
+        color_view = view_by_external_id["Color"]
+        assert color_view.referenced_containers() == {dm.ContainerId(car.BASE_MODEL.metadata.space, "Entity")}
 
 
 class TestDMSExporter: